+++ packages/patternfly-4/react-charts/src/components/ChartTheme/themes/dark/blue-color-theme.ts
@@ -1,11 +1,9 @@
 /* eslint-disable camelcase */
-import {
-  chart_color_blue_100,
-  chart_color_blue_200,
-  chart_color_blue_300,
-  chart_color_blue_400,
-  chart_color_blue_500
-} from '@patternfly/react-tokens';
+import chart_color_blue_100 from '@patternfly/react-tokens/dist/js/chart_color_blue_100';
+import chart_color_blue_200 from '@patternfly/react-tokens/dist/js/chart_color_blue_200';
+import chart_color_blue_300 from '@patternfly/react-tokens/dist/js/chart_color_blue_300';
+import chart_color_blue_400 from '@patternfly/react-tokens/dist/js/chart_color_blue_400';
+import chart_color_blue_500 from '@patternfly/react-tokens/dist/js/chart_color_blue_500';
 import { ColorTheme } from '../color-theme';
 
 // Color scale