+++ packages/patternfly-4/react-charts/src/components/ChartTheme/themes/dark/multi-color-unordered-theme.ts
@@ -1,41 +1,39 @@
 /* eslint-disable camelcase */
-import {
-  chart_color_blue_100,
-  chart_color_blue_200,
-  chart_color_blue_300,
-  chart_color_blue_400,
-  chart_color_blue_500,
-  chart_color_green_100,
-  chart_color_green_200,
-  chart_color_green_300,
-  chart_color_green_400,
-  chart_color_green_500,
-  chart_color_cyan_100,
-  chart_color_cyan_200,
-  chart_color_cyan_300,
-  chart_color_cyan_400,
-  chart_color_cyan_500,
-  chart_color_purple_100,
-  chart_color_purple_200,
-  chart_color_purple_300,
-  chart_color_purple_400,
-  chart_color_purple_500,
-  chart_color_gold_100,
-  chart_color_gold_200,
-  chart_color_gold_300,
-  chart_color_gold_400,
-  chart_color_gold_500,
-  chart_color_orange_100,
-  chart_color_orange_200,
-  chart_color_orange_300,
-  chart_color_orange_400,
-  chart_color_orange_500,
-  chart_color_black_100,
-  chart_color_black_200,
-  chart_color_black_300,
-  chart_color_black_400,
-  chart_color_black_500
-} from '@patternfly/react-tokens';
+import chart_color_blue_100 from '@patternfly/react-tokens/dist/js/chart_color_blue_100';
+import chart_color_blue_200 from '@patternfly/react-tokens/dist/js/chart_color_blue_200';
+import chart_color_blue_300 from '@patternfly/react-tokens/dist/js/chart_color_blue_300';
+import chart_color_blue_400 from '@patternfly/react-tokens/dist/js/chart_color_blue_400';
+import chart_color_blue_500 from '@patternfly/react-tokens/dist/js/chart_color_blue_500';
+import chart_color_green_100 from '@patternfly/react-tokens/dist/js/chart_color_green_100';
+import chart_color_green_200 from '@patternfly/react-tokens/dist/js/chart_color_green_200';
+import chart_color_green_300 from '@patternfly/react-tokens/dist/js/chart_color_green_300';
+import chart_color_green_400 from '@patternfly/react-tokens/dist/js/chart_color_green_400';
+import chart_color_green_500 from '@patternfly/react-tokens/dist/js/chart_color_green_500';
+import chart_color_cyan_100 from '@patternfly/react-tokens/dist/js/chart_color_cyan_100';
+import chart_color_cyan_200 from '@patternfly/react-tokens/dist/js/chart_color_cyan_200';
+import chart_color_cyan_300 from '@patternfly/react-tokens/dist/js/chart_color_cyan_300';
+import chart_color_cyan_400 from '@patternfly/react-tokens/dist/js/chart_color_cyan_400';
+import chart_color_cyan_500 from '@patternfly/react-tokens/dist/js/chart_color_cyan_500';
+import chart_color_purple_100 from '@patternfly/react-tokens/dist/js/chart_color_purple_100';
+import chart_color_purple_200 from '@patternfly/react-tokens/dist/js/chart_color_purple_200';
+import chart_color_purple_300 from '@patternfly/react-tokens/dist/js/chart_color_purple_300';
+import chart_color_purple_400 from '@patternfly/react-tokens/dist/js/chart_color_purple_400';
+import chart_color_purple_500 from '@patternfly/react-tokens/dist/js/chart_color_purple_500';
+import chart_color_gold_100 from '@patternfly/react-tokens/dist/js/chart_color_gold_100';
+import chart_color_gold_200 from '@patternfly/react-tokens/dist/js/chart_color_gold_200';
+import chart_color_gold_300 from '@patternfly/react-tokens/dist/js/chart_color_gold_300';
+import chart_color_gold_400 from '@patternfly/react-tokens/dist/js/chart_color_gold_400';
+import chart_color_gold_500 from '@patternfly/react-tokens/dist/js/chart_color_gold_500';
+import chart_color_orange_100 from '@patternfly/react-tokens/dist/js/chart_color_orange_100';
+import chart_color_orange_200 from '@patternfly/react-tokens/dist/js/chart_color_orange_200';
+import chart_color_orange_300 from '@patternfly/react-tokens/dist/js/chart_color_orange_300';
+import chart_color_orange_400 from '@patternfly/react-tokens/dist/js/chart_color_orange_400';
+import chart_color_orange_500 from '@patternfly/react-tokens/dist/js/chart_color_orange_500';
+import chart_color_black_100 from '@patternfly/react-tokens/dist/js/chart_color_black_100';
+import chart_color_black_200 from '@patternfly/react-tokens/dist/js/chart_color_black_200';
+import chart_color_black_300 from '@patternfly/react-tokens/dist/js/chart_color_black_300';
+import chart_color_black_400 from '@patternfly/react-tokens/dist/js/chart_color_black_400';
+import chart_color_black_500 from '@patternfly/react-tokens/dist/js/chart_color_black_500';
 import { ColorTheme } from '../color-theme';
 
 // The color order below improves the color contrast in unordered charts; area & line