+++ packages/patternfly-4/react-charts/src/components/ChartTheme/themes/threshold-theme.ts
@@ -1,8 +1,6 @@
 /* eslint-disable camelcase */
-import {
-  chart_threshold_stroke_dash_array,
-  chart_threshold_stroke_Width
-} from '@patternfly/react-tokens';
+import chart_threshold_stroke_dash_array from '@patternfly/react-tokens/dist/js/chart_threshold_stroke_dash_array';
+import chart_threshold_stroke_Width from '@patternfly/react-tokens/dist/js/chart_threshold_stroke_Width';
 
 // Threshold theme
 export const ThresholdTheme = {