+++ packages/patternfly-4/react-core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.136.3](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.136.2...@patternfly/react-core@3.136.3) (2020-02-04)
+
+**Note:** Version bump only for package @patternfly/react-core
+
+
+
+
+
 ## [3.136.2](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.136.1...@patternfly/react-core@3.136.2) (2020-02-03)
 
 **Note:** Version bump only for package @patternfly/react-core